[visualizer] feat: optimize the encapsulation and registration of the visualizer (#30)
* Optimize the encapsulation and registration of the visualizer

* update pr-title check
