+++ rl_insight/visualizer/__init__.py
@@ -11,3 +11,15 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+from .timeline_visualizer import RLTimelineVisualizer
+from .visualizer import (
+    BaseVisualizer,
+    get_cluster_visualizer_cls,
+)
+
+__all__ = [
+    "BaseVisualizer",
+    "get_cluster_visualizer_cls",
+    "RLTimelineVisualizer",
+]
